Move polypeptide elongation models back to process
