+++ ecoli/library/sim_data.py
@@ -87,7 +87,7 @@ class LoadSimData:
                 options in ``self.sim_data.external_state.saved_media``
             operons: Use operon transcription units
             trna_charging: Use steady-state charging model 
-                (:py:class:`~ecoli.models.polypeptide_elongation_models.SteadyStateElongationModel`)
+                (:py:class:`~ecoli.processes.polypeptide_elongation.SteadyStateElongationModel`)
                 in :py:class:`~ecoli.processes.polypeptide_elongation.PolypeptideElongation`
             ppgpp_regulation: Enable growth rate control using ppGpp in 
                 polypeptide elongation and transcript initiation
@@ -136,7 +136,7 @@ class LoadSimData:
             mechanistic_aa_transport: Constrain amino acid uptake based on 
                 external concentrations and exchange rates in 
                 :py:class:`~ecoli.processes.metabolism.Metabolism`
-            translation_supply: Use :py:class:`~ecoli.models.polypeptide_elongation_models.TranslationSupplyElongationModel` 
+            translation_supply: Use :py:class:`~ecoli.processes.polypeptide_elongation.TranslationSupplyElongationModel` 
                 in :py:class:`~ecoli.processes.polypeptide_elongation.PolypeptideElongation`. 
                 Superseded by ``trna_charging``
             aa_supply_in_charging: Calculate charged tRNA supply from each sub
